Create controlled-value.json
A new FDO type to accommodate the controlled values in terms. Example topicDiscipline has a value called Botany. We need a schema to indicate that this a value and it belongs to a term and also vice versa. In essence the term and  controlled-value are mostly same but with the above distinction. we need separate entities to link terms with values and extend the term also in future have hierarchies.